Doctor Consultation Portal/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.36
pymysql==1.1.0
cryptography>=41.0.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.0.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0

